rtl8xxxu: Add some 8188eu registers and update CCK0_AFE_SETTING bit defines
CCK0_AFE_SETTING is particular, it has the notion of primary RX antenna
and optional RX antenna. When configuring RX for single antenna, setup
should use the same antenna for default and optional. For AB setup,
use antenna A as default and B as optional.

In addition add info for 8188eu IOL magic interface used to send
firmware and register init files to the firmware.

Signed-off-by: Jes Sorensen <Jes.Sorensen@redhat.com>